+++ NYTimesSampleApp/Extensions/UIExtensions/UIViewControllerExtensions.swift
@@ -9,7 +9,7 @@
 import UIKit
 
 protocol Instantiatable {
-    static func instantiateInitalViewController(from storyboardIdentifier: UIStoryboard.Storyboard) -> Self
+    static func instantiateInitialViewController(from storyboardIdentifier: UIStoryboard.Storyboard) -> Self
     static func instantiate(from storyboardIdentifier: UIStoryboard.Storyboard) -> Self
 }
 
@@ -17,7 +17,7 @@ extension UIViewController: Instantiatable { }
 
 extension Instantiatable where Self: UIViewController {
 
-    static func instantiateInitalViewController(from storyboardIdentifier: UIStoryboard.Storyboard) -> Self {
+    static func instantiateInitialViewController(from storyboardIdentifier: UIStoryboard.Storyboard) -> Self {
 
         guard let viewController = UIStoryboard(storyboard: storyboardIdentifier).instantiateInitialViewController() as? Self else {
             return Self()
