+++ NYTimesSampleApp/SupportingFiles/Resources/String/en.lproj/Localizable.strings
@@ -1,0 +1,12 @@
+/* 
+  Localizable.strings
+  NYTimesSampleApp
+
+  Created by Ahmad Jabri on 27/11/2020.
+  Copyright © 2020 Ahmad Jabri. All rights reserved.
+*/
+
+// MARK: - Common
+"Common.Welcome" = "Welocme";
+"Common.NYTimes.Title" = "NY Time Most Popular";
+
